14.14.0 RPM final - round 0

+++ RPM/scripts/postinst/kaltura-config-all.sh
@@ -127,6 +127,13 @@ echo "Running DWH config...
 $BASE_DIR/bin/kaltura-dwh-config.sh "$ANSFILE" 
 if [ $? -ne 0 ];then
        echo -e "${BRIGHT_RED}ERROR: $BASE_DIR/bin/kaltura-dwh-config.sh failed:( You can re-run it when the issue is fixed.${NORMAL}"
+	exit 115
+fi
+
+$BASE_DIR/bin/kaltura-es-config.sh "$ANSFILE"
+if [ $? -ne 0 ];then
+       echo -e "${BRIGHT_RED}ERROR: $BASE_DIR/bin/kaltura-es-config.sh failed:( You can re-run it when the issue is fixed.${NORMAL}"
+	exit 117
 fi
 find $APP_DIR/cache/ -type f -exec rm {} \;
 rm -f $APP_DIR/log/kaltura-*.log
